Attach cryptographic identity to each entity.
We also sign our public key with our owners private key and keep the
signature around so we can verify that we came from that owner.

This is also useful for a few other reasons, like being able to store
data we can later verify came from "us" I also imagine being able to
create identities and store them on batches of entities.