Rewrite serialization approach, fix some wonky kwargs stuff.
This was a wild ride... using jsonpickle now and relying on that to
actually produce a hierarchy with instances. This involved some magic
trickery to fool jsonpickle into serializing the full root class but
then only references on other instances of that class in the
hieararchy.

Web is very broken.