More serialization tweaking, getting way better.
This introduces wrapt to proxy the entity refs, instead of the dict
tricks I was pulling earlier.

The idea here is to deserialize entities as EntityRef's which are
proxy objects and then we fixup the __wrapped__ target attribute
afterwards. At first I thought we may have to store EntityRef's in the
collections, but we can still avoid that step, for now.