feat: Foundation for Chimeric entities.
This is the beginning of a large refactor to pull apart the larger
Entities like Area and Things that have tons of mixins into more
silo'd Python modules.

Each entity will eventually have separate transactional scopes, each
defined by a particular behavior. These line up very well with some of
the mixins. So an entity in the database has multiple rows, but
identity wise it's still just a single thing. This isn't the case, yet
but this is a step.

So let's call these scopes or whatever.

`holding`
`health`
`navigable`

These are separate things.

This makes for a much better separate of data and the behavior that we
were seeing with the move to push everything into props.

So in this system each scope has props and scopes can include each
other so they can share. So we can have a core scope and then holding
scope.

Each scope has a grammar and all the associted interaction.