fix: Maintain global scope across dynamic calls.
Not sure if this is the best way to go about this but the general idea
is that after the initial eval to bind handlers any supporting
declarations end up in the locals dict passed to eval. These then need
to be available when we call the handler, so thta means a call to eval
where we can override the globals on. The alternative is to merge into
our global namespace, which would suck.