feat: Decouple usernames from keys.
Notice that the map is stored in the world using a scope. This means
if servers ever get large enough the map will need to be handled much
better. I think that would be a good time to evolve some nifty
bindings backed by sqlite or something for doing lazy lookups by key.