feat: Generate check from json diffs on saves.
I really like this idea as a solution to a few problems. Basically,
we're walking the diff of an update and finding Acls along the way,
those are then the ones we'll require pass muster.

We're basically locked into the json diff thing, which I'm fine with
at this point because it's solving a ton of issues.