feat: Migration facility.
Right now this just resaves all the entities in the database after
loading them, which if we're lucky will be the only real migration
we'll need. I have a feeling that won't hold, though.

I imagine we'll eventually need to be able to run python code as part
of this somehow.